Thinh Update 21/05/224
