+++ shopapp-angular/src/app/components/order/order.component.html
@@ -130,10 +130,10 @@
                 </table>
                </div>
                 <div class="text-start mt-3 coupon-and-price-container">
-                    <div class="input-group">
-                        <input type="text" class="form-controls coupon" formControlName="couponCode" placeholder="Enter coupon">
-                        <button (click)="applyCoupon()" class="btn btn-light" type="button">Apply</button>
-                    </div>
+<!--                    <div class="input-group">-->
+<!--                        <input type="text" class="form-controls coupon" formControlName="couponCode" placeholder="Enter coupon">-->
+<!--                        <button (click)="applyCoupon()" class="btn btn-light" type="button">Apply</button>-->
+<!--                    </div>-->
                     <h4 class="header-text text-end   total-price">Total price: ${{ totalAmount | number:'1.2-2'}} </h4>
                 </div>
