+++ shopapp-backend/src/main/java/ChromeBrowser/WebUtils.java
@@ -1,0 +1,11 @@
+package ChromeBrowser;
+
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
+
+public class WebUtils {
+    public static HttpServletRequest getCurrentRequest() {
+        return ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest();
+    }
+}
