--- shopapp-backend/src/main/java/com/project/shopapp/utils/WebUtils.java
@@ -1,11 +1,0 @@
-package com.project.shopapp.utils;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-public class WebUtils {
-    public static HttpServletRequest getCurrentRequest() {
-        return ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest();
-    }
-}
